Release 3.3.6 (#285)
Add ID 2.5 statement approval functionality
Introduce ConcordiumClient wrapper over SDK
Moved Lottie animations from DotLottie on Lottie framework
Update Reown(Wallet Connect) flow UI + alerts logic
Add new type request_verifiable_presentation for incoming verify transactions type
