+++ ConcordiumWallet/Features/WalletConnect/SessionRequestFlow/SessionRequestDataType.swift
@@ -6,36 +6,9 @@
 //  Copyright © 2024 pioneeringtechventures. All rights reserved.
 //
 
-import Foundation
-import ReownWalletKit
-
-struct SessionRequestType: Codable {
-    let type: TransferType
-    
-    enum CodingKeys: String, CodingKey {
-        case type
-    }
-
-    init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        
-        // Decode 'type' field into TransferType.
-        let typeStr = try container.decode(String.self, forKey: .type)
-        if typeStr == "Update" {
-            // For backwards compatibility with older versions of @concordium/wallet-connectors.
-            type = TransferType.transferUpdate
-        } else if typeStr == "transfer" {
-            type = TransferType.simpleTransfer
-        } else if let t = TransferType(rawValue: typeStr) {
-            type = t
-        } else {
-            throw DecodingError.dataCorruptedError(forKey: .type, in: container, debugDescription: "Invalid transaction type '\(typeStr)'")
-        }
-    }
-}
-
 import Concordium
 import Foundation
+import ReownWalletKit
 
 enum SessionRequestDataType {
     case signMessage(SignMessagePayload)
@@ -43,25 +16,17 @@ enum SessionRequestDataType {
     case signAndSend(ContractUpdateRequestParams)
     case verifiablePresentation(WalletConnectRequestVerifiablePresentationParam)
     
-    /// Basically we support two types of incoming wallet connect requests: `sign_message` & `sign_message`
-    /// In case of `sign_message` wee need to determine what king of request is: send or update.
-    /// having this info we can properly map income request data payload
     init(sessionRequest: Request) throws {
         switch sessionRequest.method {
             case "request_verifiable_presentation":
                 do {
-                    struct DummyJSON: Codable {
+                    struct ParamsData: Codable {
                         let paramsJson: String
                     }
-                    let dummy = try sessionRequest.params.get(DummyJSON.self)
-//                    let payload: VerifiablePresentationRequestParams = try JSONDecoder().decode(VerifiablePresentationRequestParams.self, from: Data(dummy.paramsJson.utf8))
-//                    self = .verifiablePresentation(payload)
-
-                    
+                    let paramsData = try sessionRequest.params.get(ParamsData.self)
                     let formatter = ISO8601DateFormatter()
                     formatter.formatOptions = [.withInternetDateTime, .withFractionalSeconds]
-                    let value = try JSONDecoder().decode(WalletConnectRequestVerifiablePresentationParam.self, from: dummy.paramsJson.data(using: .utf8)!)
-                    
+                    let value = try JSONDecoder().decode(WalletConnectRequestVerifiablePresentationParam.self, from: paramsData.paramsJson.data(using: .utf8)!)
                     
                     self = .verifiablePresentation(value)
                 } catch {
